+++ crates/blockchain/mempool.rs
@@ -130,7 +130,8 @@ impl Mempool {
         &self,
         filter: &dyn Fn(&Transaction) -> bool,
     ) -> Result<HashMap<Address, Vec<MempoolTransaction>>, StoreError> {
-        let mut txs_by_sender: HashMap<Address, Vec<MempoolTransaction>> = HashMap::new();
+        let mut txs_by_sender: HashMap<Address, Vec<MempoolTransaction>> =
+            HashMap::with_capacity(128);
         let tx_pool = self
             .transaction_pool
             .read()
@@ -140,7 +141,7 @@ impl Mempool {
             if filter(tx) {
                 txs_by_sender
                     .entry(tx.sender())
-                    .or_default()
+                    .or_insert_with(|| Vec::with_capacity(128))
                     .push(tx.clone())
             }
         }
@@ -176,7 +177,7 @@ impl Mempool {
             .read()
             .map_err(|error| StoreError::MempoolReadLock(error.to_string()))?
             .get(&transaction_hash)
-            .map(|e| e.clone().into());
+            .map(|e| e.transaction().clone());
 
         Ok(tx)
     }